+++ music-api/src/main/java/music/exception/LibraryNotFoundException.kt
@@ -1,0 +1,3 @@
+package music.exception
+
+class LibraryNotFoundException(libraryId: Long) : Exception("Library ${libraryId}does not exist.")