+++ music-api/src/main/java/music/model/SyncResult.kt
@@ -1,10 +1,10 @@
 package music.model
 
 class SyncResult(
-        var newTracks: List<DeferredTrack>,
-        var unmodifiedTracks: List<DeferredTrack>,
-        var modifiedTracks: List<DeferredTrack>,
-        var failedTracks: List<DeferredTrack>,
-        var orphanedTracks: List<Track>,
-        var unorphanedTracks: List<Track>
+	var newTracks: List<DeferredTrack> = ArrayList(),
+	var unmodifiedTracks: List<DeferredTrack> = ArrayList(),
+	var modifiedTracks: List<DeferredTrack> = ArrayList(),
+	var failedTracks: List<DeferredTrack> = ArrayList(),
+	var orphanedTracks: List<Track> = ArrayList(),
+	var unorphanedTracks: List<Track> = ArrayList()
 )