+++ music-api/src/main/java/music/repository/ILibraryRepository.kt
@@ -1,0 +1,9 @@
+package music.repository
+
+import music.model.Library
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ILibraryRepository : JpaRepository<Library, Long>{
+}