+++ music-api/src/main/java/music/service/ConvertService.kt
@@ -8,7 +8,6 @@ import org.apache.commons.codec.digest.DigestUtils
 import org.apache.commons.exec.CommandLine
 import org.apache.commons.exec.DefaultExecutor
 import org.apache.commons.io.FileUtils
-import org.apache.commons.io.FilenameUtils
 import org.apache.commons.lang3.StringUtils
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -31,16 +30,16 @@ class ConvertService {
 
 	fun convertFile(device: Device, track: Track): ByteArray? {
 		try {
-			logger.trace("Converting {}", track.location)
+			logger.trace("Converting {}", track.libraryPath)
 
 			/*
 			ffmpeg does not like spaces in the source directory path, and I have tried numerous ways of fixing it to
 			no avail. Instead, it seems much simpler to just copy the source file to the temp file location, which
 			usually ensures that it won't have a space in it. At least on Linux, which is what I really care about.
 			 */
-			val sourceTemp = File.createTempFile("source", ".${FilenameUtils.getExtension(track.location)}")
+			val sourceTemp = File.createTempFile("source", ".${track.extension}")
 			sourceTemp.deleteOnExit()
-			FileUtils.copyFile(fileService.getFile(track.location), sourceTemp)
+			FileUtils.copyFile(fileService.getFile(track.libraryPath), sourceTemp)
 
 			// todo, figure out a way to convert the file into memory rather than on disk (likely not possible due to using ffmpeg)
 			val target = File.createTempFile("example", ".${device.format}")
@@ -81,10 +80,10 @@ class ConvertService {
 				FileUtils.deleteQuietly(target)
 				return fileBytes
 			} else {
-				throw Exception("Failed to convert ${track.location} with ffmpeg exit value of $exitValue")
+				throw Exception("Failed to convert ${track.libraryPath} with ffmpeg exit value of $exitValue")
 			}
 		} catch (ex: Exception) {
-			logger.error("Failed to convert ${track.location}", ex)
+			logger.error("Failed to convert ${track.libraryPath}", ex)
 			return null
 		}
 	}