+++ music-api/src/main/java/music/service/SyncService.kt
@@ -3,6 +3,7 @@ package music.service
 import music.endpoint.AdminEndpoint
 import music.exception.TaskInProgressException
 import music.model.SyncResult
+import music.repository.ILibraryRepository
 import org.jaudiotagger.audio.exceptions.CannotReadException
 import org.jaudiotagger.audio.exceptions.InvalidAudioFrameException
 import org.jaudiotagger.audio.exceptions.ReadOnlyFileException
@@ -11,31 +12,32 @@ import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import java.io.IOException
-import java.util.ArrayList
 import java.util.concurrent.atomic.AtomicBoolean
 
 @Service
-class SyncService @Autowired constructor(private val metadataService: MetadataService, private val trackService: TrackService) {
+class SyncService @Autowired constructor(private val metadataService: MetadataService,
+										 private val trackService: TrackService,
+										 private val libraryRepository: ILibraryRepository) {
 
 	private val logger = LoggerFactory.getLogger(AdminEndpoint::class.java)
-	private val currentlySyncing: AtomicBoolean = AtomicBoolean(false);
+	private val currentlySyncing: AtomicBoolean = AtomicBoolean(false)
+
 
 	@Throws(ReadOnlyFileException::class, CannotReadException::class, TagException::class, InvalidAudioFrameException::class, IOException::class)
-	fun syncTracksToDb(forceUpdates: Boolean = false): SyncResult {
+	fun syncTracksToDb(forceUpdates: Boolean = false, libraryId: Long): SyncResult {
 		checkNotSyncing()
 		try {
 			logger.info("Begin database sync")
 			currentlySyncing.set(true)
-			val syncResult = SyncResult(ArrayList(), ArrayList(), ArrayList(), ArrayList(), ArrayList(), ArrayList())
-			val files = metadataService.getTracks()
+			val syncResult = SyncResult()
+			val library = libraryRepository.findById(libraryId).get()
+			val files = metadataService.getTracks(library)
 			trackService.upsertTracks(files, syncResult, forceUpdates)
-			trackService.deleteOrphanedTracksMetadata(files, syncResult)
+			trackService.deleteOrphanedTracksMetadata(files, syncResult, library)
 			logger.info("Finished database sync")
-			currentlySyncing.set(false)
 			return syncResult
-		} catch (e: Exception) {
+		} finally {
 			currentlySyncing.set(false)
-			throw e
 		}
 	}